+++ floris/turbine_library/__init__.py
@@ -1,1 +1,5 @@
 from floris.turbine_library.turbine_previewer import TurbineInterface, TurbineLibrary
+from floris.turbine_library.turbine_utilities import (
+    build_cosine_loss_turbine_dict,
+    check_smooth_power_curve,
+)
